deployment/requirements.txt
mcp>=1.0.0
playwright>=1.40.0
keyring>=24.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
